+++ clang/include/joe_String.h
@@ -27,7 +27,7 @@ JOEOBJ_API joe_Object joe_String_New2 (const char *c1, const char *c2);
 JOEOBJ_API joe_Object joe_String_New3 (const char *c1, const char *c2, const char *c3);
 JOEOBJ_API joe_Object joe_String_New4 (const char *c1, const char *c2, const char *c3,
                             const char *c4);
-JOEOBJ_API joe_Object joe_String_New_len (const char *c, unsigned int len);
+JOEOBJ_API joe_Object joe_String_New_len (const char *c, size_t len);
 JOEOBJ_API char * joe_String_getCharStar (joe_String self);
 JOEOBJ_API unsigned int joe_String_length (joe_String self);
 JOEOBJ_API int joe_String_compare (joe_String self, joe_String aStr);
